Updates for LJG proposal

+++ conversion_utils.py
@@ -79,7 +79,7 @@ def beamgridder(xcen,ycen,size):
     if round(ycen) > size - 1 or round(xcen) > size - 1 or ycen < 0. or xcen <0.:
         return beam
     else:
-        beam[round(ycen),round(xcen)] = 1. #single pixel gridder
+        beam[int(round(ycen)), int(round(xcen))] = 1. #single pixel gridder
         return beam
 
 def closest(arr, val):
